refactor: consolidate requirements, remove dead code, fix notebook deps
