--- requirements-colab.txt
@@ -1,29 +1,0 @@
-pytest==8.3.3
-pytest-asyncio==0.24.0
-ruff==0.6.8
-pydantic==2.9.2
-pyyaml==6.0.2
-python-dotenv==1.0.1
-aiohttp==3.13.3
-asyncpraw==7.8.1
-asyncprawcore==2.4.0
-moviepy==2.1.1
-numpy==2.1.2
-opencv-python==4.10.0.84
-psutil==6.0.0
-yt-dlp>=2024.0
-faster-whisper==1.0.0
-aiosqlite==0.20.0
-openai
-nest_asyncio
-google-api-python-client
-google-auth-oauthlib
-google-auth-httplib2
-opencv-python-headless
-pillow
-scipy
-librosa
-soundfile
-gputil
-filelock
-pynvml
