+++ requirements.txt
@@ -23,6 +23,12 @@ filelock>=3.13
 chromadb>=1.5.0
 sentence-transformers>=2.2.0
 
-# Optional: torch, transformers (uncomment if needed)
-# torch>=2.0.0
-# transformers>=4.30.0
+torch>=2.0.0
+transformers>=4.30.0
+
+# Colab
+nest_asyncio
+opencv-python-headless
+gputil
+pynvml
+asyncprawcore>=2.4.0
