+++ src/hybrid_documentary_state_machine.py
@@ -9,7 +9,6 @@ import logging
 import os
 import re
 import time
-from dataclasses import dataclass
 from datetime import UTC, datetime
 from enum import Enum
 from pathlib import Path
@@ -26,11 +25,9 @@ from typing import (
 )
 from uuid import uuid4
 
-import aiohttp
-
-if TYPE_CHECKING:
-    from src.utils.search_audit_logger import SearchAuditLogger
 from pydantic import BaseModel, ConfigDict, Field, ValidationError
+
+from src.integrations.search_client import ExaSearchClient
 
 OpenAI = None
 try:
@@ -148,15 +145,6 @@ class RunState(BaseModel):
     metadata: Dict[str, Any] = Field(default_factory=dict)
 
     model_config = ConfigDict(validate_assignment=True)
-
-
-@dataclass(frozen=True)
-class MediaSearchResult:
-    title: str
-    url: str
-    description: str
-    source: str
-    thumbnail_url: str = ""
 
 
 class SummaryResult(BaseModel):
@@ -434,151 +422,6 @@ def save_finding(
     return target_path
 
 
-class ExaSearchClient:
-    """Exa web search client via Hack Club AI proxy."""
-
-    def __init__(
-        self,
-        api_key: Optional[str] = None,
-        base_url: str = "https://ai.hackclub.com/proxy/v1/exa",
-        timeout_seconds: int = 20,
-        audit_logger: Optional["SearchAuditLogger"] = None,
-    ):
-        self.api_key = api_key or os.getenv("HACKCLUB_SEARCH_API_KEY") or ""
-        self.base_url = base_url.rstrip("/")
-        self.timeout_seconds = timeout_seconds
-        self.audit_logger = audit_logger
-        self._session: Optional[aiohttp.ClientSession] = None
-        self._session_lock = asyncio.Lock()
-
-    async def _ensure_session(self) -> aiohttp.ClientSession:
-        if self._session is not None and not self._session.closed:
-            return self._session
-        async with self._session_lock:
-            if self._session is None or self._session.closed:
-                self._session = aiohttp.ClientSession(
-                    timeout=aiohttp.ClientTimeout(total=self.timeout_seconds)
-                )
-        return self._session
-
-    async def close(self) -> None:
-        if self._session is not None and not self._session.closed:
-            await self._session.close()
-        self._session = None
-
-    async def __aenter__(self) -> "ExaSearchClient":
-        await self._ensure_session()
-        return self
-
-    async def __aexit__(self, exc_type, exc, tb) -> None:  # type: ignore[override]
-        await self.close()
-
-    async def search_web(self, query: str, count: int = 10) -> List[MediaSearchResult]:
-        return await self._search(query, count)
-
-    async def search_images(
-        self, query: str, count: int = 10
-    ) -> List[MediaSearchResult]:
-        logger.warning(
-            "Exa does not support image search; use BraveImageClient instead"
-        )
-        return []
-
-    async def search_videos(
-        self, query: str, count: int = 10
-    ) -> List[MediaSearchResult]:
-        logger.warning("Exa does not support video search; use Brave directly")
-        return []
-
-    async def _search(
-        self,
-        query: str,
-        count: int,
-    ) -> List[MediaSearchResult]:
-        if not self.api_key:
-            logger.warning("HACKCLUB_SEARCH_API_KEY is not configured")
-            return []
-
-        session = await self._ensure_session()
-        headers = {
-            "Authorization": f"Bearer {self.api_key}",
-            "Content-Type": "application/json",
-        }
-        body = {"query": query, "numResults": max(1, min(count, 50))}
-        url = f"{self.base_url}/search"
-
-        if self.audit_logger:
-            self.audit_logger.log_request("POST", url, body, headers)
-
-        start = time.perf_counter()
-        try:
-            async with session.post(url, headers=headers, json=body) as response:
-                resp_body = await response.text()
-                duration_ms = (time.perf_counter() - start) * 1000
-                if self.audit_logger:
-                    self.audit_logger.log_response(
-                        response.status,
-                        resp_body,
-                        duration_ms,
-                        url=url,
-                    )
-                if response.status >= 400:
-                    logger.warning(
-                        "Exa search failed status=%s body=%s",
-                        response.status,
-                        resp_body[:300],
-                    )
-                    return []
-                try:
-                    payload = json.loads(resp_body)
-                except json.JSONDecodeError:
-                    return []
-        except Exception as exc:
-            duration_ms = (time.perf_counter() - start) * 1000
-            if self.audit_logger:
-                self.audit_logger.log_response(500, str(exc), duration_ms, url=url)
-            logger.error("Exa search request failed: %s", exc)
-            return []
-
-        return self._parse_search_results(payload)
-
-    def _parse_search_results(
-        self,
-        payload: Any,
-    ) -> List[MediaSearchResult]:
-        candidates: List[Dict[str, Any]] = []
-
-        if isinstance(payload, list):
-            candidates = [item for item in payload if isinstance(item, dict)]
-        elif isinstance(payload, dict):
-            candidates = [
-                item for item in payload.get("results", []) if isinstance(item, dict)
-            ]
-
-        parsed: List[MediaSearchResult] = []
-        for item in candidates:
-            url = item.get("url") or item.get("link") or ""
-            if not isinstance(url, str) or not url:
-                continue
-
-            title = item.get("title") or ""
-            description = item.get("text") or item.get("snippet") or ""
-            source = item.get("source") or "unknown"
-
-            parsed.append(
-                MediaSearchResult(
-                    title=str(title),
-                    url=url,
-                    description=str(description),
-                    source=str(source),
-                    thumbnail_url="",
-                )
-            )
-
-        logger.debug("Parsed %d Exa search results", len(parsed))
-        return parsed
-
-
 def transcribe_media_file(
     media_path: Union[str, Path],
     project_dir: Union[str, Path],
@@ -813,8 +656,6 @@ def build_state_machine_prompt(
 __all__ = [
     "DEFAULT_SUMMARY_MODEL",
     "DEFAULT_TRANSCRIBE_MODEL",
-    "ExaSearchClient",
-    "MediaSearchResult",
     "NVIDIA_BASE_URL",
     "PHASE_JSON_CONTRACTS",
     "PHASE_VALIDATION_SCHEMAS",
