+++ src/integrations/tts_service.py
@@ -179,12 +179,6 @@ class TTSService:
             if result:
                 return result
 
-        try:
-            pyttsx3_module = importlib.import_module("pyttsx3")
-            return self._generate_with_pyttsx3(segment, output_path, pyttsx3_module)
-        except ModuleNotFoundError:
-            pass
-
         self.logger.error("No TTS service available for speech generation")
         return None
 
@@ -633,31 +627,3 @@ class TTSService:
         if QWEN_TTS_AVAILABLE:
             services.append("qwen3_tts")
         return services
-
-    def _generate_with_pyttsx3(
-        self, segment: NarrativeSegment, output_path: Path, pyttsx3_module: Any
-    ) -> Optional[Path]:
-        """Generate speech using system TTS (pyttsx3) as fallback."""
-        try:
-            engine = pyttsx3_module.init()
-
-            # Configure voice settings based on emotion and pacing
-            rate = 200  # Default rate
-            if segment.pacing == PacingType.SLOW:
-                rate = 150
-            elif segment.pacing == PacingType.FAST:
-                rate = 250
-
-            engine.setProperty("rate", rate)
-            engine.setProperty("volume", 0.9)
-
-            # Save to file
-            engine.save_to_file(segment.text, str(output_path))
-            engine.runAndWait()
-
-            self.logger.info(f"Generated pyttsx3 TTS: '{segment.text[:30]}...'")
-            return output_path
-
-        except Exception as e:
-            self.logger.warning(f"pyttsx3 TTS failed: {e}")
-            return None
