--- src/processing/enhancement_optimizer.py
@@ -1,922 +1,0 @@
-"""
-Enhancement Optimizer - Automated Performance Optimization System
-Analyzes video performance data and automatically fine-tunes enhancement parameters.
-"""
-
-import logging
-import json
-import os
-import tempfile
-import numpy as np
-import yaml
-from pathlib import Path
-from typing import Dict, List, Optional, Tuple, Any
-from datetime import datetime, timedelta
-from dataclasses import dataclass, asdict
-import statistics
-
-from src.config.settings import get_config
-from src.models import EnhancementOptimization, PerformanceMetrics
-from src.monitoring.engagement_metrics import EngagementAnalyzer, EngagementMetricsDB
-
-
-@dataclass
-class OptimizationRecommendation:
-    """Recommendation for parameter adjustment"""
-
-    parameter_name: str
-    current_value: float
-    recommended_value: float
-    confidence: float  # 0.0 to 1.0
-    impact_estimate: float  # Expected improvement percentage
-    reason: str
-
-
-class EnhancementOptimizer:
-    """
-    Automated system that analyzes video performance and optimizes
-    enhancement parameters for better engagement and retention.
-    """
-
-    def __init__(self):
-        self.config = get_config()
-        self.logger = logging.getLogger(__name__)
-        self.db = EngagementMetricsDB()
-        self.analyzer = EngagementAnalyzer(self.db)
-
-        # Optimization parameters
-        self.min_sample_size = 10  # Minimum videos for statistical significance
-        self.confidence_threshold = 0.7  # Minimum confidence for auto-adjustment
-        self.max_adjustment_per_cycle = 0.2  # Maximum change per optimization cycle
-        self.optimization_cycle_days = 7  # Days between optimization cycles
-
-        # Parameter ranges and constraints
-        self.parameter_constraints = {
-            # Only include parameters that are wired into the runtime config.
-            "zoom_intensity": {"min": 1.0, "max": 1.8, "step": 0.1},
-            "color_grading_strength": {"min": 0.0, "max": 1.0, "step": 0.05},
-            "background_music_volume": {"min": 0.1, "max": 0.8, "step": 0.05},
-        }
-
-        self.parameter_yaml_paths = {
-            "background_music_volume": ("audio", "background_music", "volume"),
-            "zoom_intensity": ("effects", "max_zoom"),
-            "color_grading_strength": ("effects", "color_grade_intensity"),
-        }
-
-        # Load current optimization state
-        self.optimization_state = self._load_optimization_state()
-
-    def optimize_parameters(self, force_optimization: bool = False) -> Dict[str, Any]:
-        """
-        Run comprehensive parameter optimization based on performance data
-
-        Args:
-            force_optimization: Force optimization even if cycle hasn't completed
-
-        Returns:
-            Optimization results and recommendations
-        """
-        try:
-            self.logger.info("Starting enhancement parameter optimization...")
-
-            # Check if optimization cycle is due
-            if not force_optimization and not self._is_optimization_due():
-                return {"status": "skipped", "reason": "optimization_cycle_not_due"}
-
-            # Analyze recent performance data
-            performance_analysis = self._analyze_recent_performance()
-
-            if not performance_analysis["sufficient_data"]:
-                return {
-                    "status": "insufficient_data",
-                    "min_required": self.min_sample_size,
-                }
-
-            # Generate optimization recommendations
-            recommendations = self._generate_optimization_recommendations(
-                performance_analysis
-            )
-
-            # Apply high-confidence recommendations automatically
-            applied_changes = self._apply_auto_optimizations(recommendations)
-
-            # Update optimization state
-            self._update_optimization_state(recommendations, applied_changes)
-
-            # Generate optimization report
-            report = self._generate_optimization_report(
-                performance_analysis, recommendations, applied_changes
-            )
-
-            self.logger.info(
-                f"Optimization complete: {len(applied_changes)} changes applied"
-            )
-            return report
-
-        except Exception as e:
-            self.logger.error(f"Parameter optimization failed: {e}")
-            return {"status": "error", "error": str(e)}
-
-    def _is_optimization_due(self) -> bool:
-        """Check if optimization cycle is due"""
-        last_optimization = self.optimization_state.get("last_optimization")
-        if not last_optimization:
-            return True
-
-        last_date = datetime.fromisoformat(last_optimization)
-        days_since = (datetime.now() - last_date).days
-
-        return days_since >= self.optimization_cycle_days
-
-    def _analyze_recent_performance(self) -> Dict[str, Any]:
-        """Analyze recent video performance data"""
-        try:
-            analysis_period = 30  # Days
-
-            # Get performance data for different enhancement types
-            enhancement_performance = {}
-            total_videos = 0
-
-            for enhancement_type in [
-                "sound_effects",
-                "visual_effects",
-                "text_overlays",
-                "color_grading",
-                "dynamic_zoom",
-                "speed_effects",
-            ]:
-                performance = self.analyzer.analyze_enhancement_performance(
-                    enhancement_type, analysis_period
-                )
-
-                if performance:
-                    enhancement_performance[enhancement_type] = asdict(performance)
-                    total_videos = max(
-                        total_videos,
-                        performance.videos_with_enhancement
-                        + performance.videos_without_enhancement,
-                    )
-
-            # Calculate parameter-specific correlations
-            parameter_correlations = self._calculate_parameter_correlations()
-
-            # Determine if we have sufficient data
-            sufficient_data = total_videos >= self.min_sample_size
-
-            analysis = {
-                "analysis_period_days": analysis_period,
-                "total_videos_analyzed": total_videos,
-                "sufficient_data": sufficient_data,
-                "enhancement_performance": enhancement_performance,
-                "parameter_correlations": parameter_correlations,
-                "baseline_metrics": self._calculate_baseline_metrics(),
-                "performance_trends": self._analyze_performance_trends(),
-            }
-
-            return analysis
-
-        except Exception as e:
-            self.logger.error(f"Performance analysis failed: {e}")
-            return {"sufficient_data": False, "error": str(e)}
-
-    def _calculate_parameter_correlations(self) -> Dict[str, Dict[str, float]]:
-        """Calculate actual correlations from stored performance data.
-
-        TODO: This is a stub. The real implementation should:
-        1. Query the engagement_metrics database for parameter values and
-           corresponding performance metrics (engagement_rate, retention_rate,
-           completion_rate) over the last 30 days.
-        2. Compute Pearson/Spearman correlations between each parameter and
-           each metric using numpy or scipy.
-        3. Only return correlations when sufficient data exists (n >= 10 videos).
-
-        Until enough real data is collected, returns heuristic defaults.
-        """
-        correlations: Dict[str, Dict[str, float]] = {}
-
-        try:
-            with sqlite3.connect(str(self.db.db_path)) as conn:
-                cursor = conn.cursor()
-                cursor.execute(
-                    """
-                    SELECT enhancements_used, engagement_rate, retention_rate,
-                           completion_rate, click_through_rate
-                    FROM video_metrics
-                    WHERE upload_date >= ?
-                    """,
-                    ((datetime.now() - timedelta(days=30)).isoformat(),),
-                )
-                rows = cursor.fetchall()
-                if len(rows) >= self.min_sample_size:
-                    self.logger.info(
-                        "Computing parameter correlations from %d videos", len(rows)
-                    )
-                    for param in self.parameter_constraints:
-                        param_values = []
-                        eng_values = []
-                        ret_values = []
-                        cmp_values = []
-                        for row in rows:
-                            enhancements = json.loads(row[0]) if row[0] else []
-                            if param in enhancements:
-                                param_values.append(1.0)
-                            else:
-                                param_values.append(0.0)
-                            eng_values.append(row[1] or 0.0)
-                            ret_values.append(row[2] or 0.0)
-                            cmp_values.append(row[3] or 0.0)
-                        if len(set(param_values)) > 1:
-                            eng_corr = (
-                                np.corrcoef(param_values, eng_values)[0, 1]
-                                if np.std(param_values) > 0
-                                else 0
-                            )
-                            ret_corr = (
-                                np.corrcoef(param_values, ret_values)[0, 1]
-                                if np.std(param_values) > 0
-                                else 0
-                            )
-                            cmp_corr = (
-                                np.corrcoef(param_values, cmp_values)[0, 1]
-                                if np.std(param_values) > 0
-                                else 0
-                            )
-                            correlations[param] = {
-                                "engagement_correlation": float(eng_corr),
-                                "retention_correlation": float(ret_corr),
-                                "completion_correlation": float(cmp_corr),
-                            }
-                    return correlations
-                self.logger.info(
-                    "Insufficient data for correlation analysis (%d videos, need %d)",
-                    len(rows),
-                    self.min_sample_size,
-                )
-        except Exception as e:
-            self.logger.warning("Parameter correlation calculation failed: %s", e)
-
-        return correlations
-
-    def _calculate_baseline_metrics(self) -> Dict[str, float]:
-        """Calculate baseline performance metrics from real data.
-
-        Attempts to query the engagement_metrics database for average metrics
-        across all recent videos. Falls back to empty dict if insufficient data.
-        """
-        try:
-            with sqlite3.connect(str(self.db.db_path)) as conn:
-                cursor = conn.cursor()
-                cursor.execute(
-                    """
-                    SELECT AVG(engagement_rate), AVG(retention_rate),
-                           AVG(completion_rate), AVG(click_through_rate)
-                    FROM video_metrics
-                    WHERE upload_date >= ?
-                    """,
-                    ((datetime.now() - timedelta(days=30)).isoformat(),),
-                )
-                row = cursor.fetchone()
-                if row and row[0] is not None:
-                    return {
-                        "avg_engagement_rate": row[0] or 0.0,
-                        "avg_retention_rate": row[1] or 0.0,
-                        "avg_completion_rate": row[2] or 0.0,
-                        "avg_ctr": row[3] or 0.0,
-                    }
-        except Exception as e:
-            self.logger.warning("Baseline calculation failed: %s", e)
-
-        return {}
-
-    def _analyze_performance_trends(self) -> Dict[str, Any]:
-        """Analyze performance trends over time from real DB data."""
-        try:
-            trends: Dict[str, Any] = {
-                "engagement_trend": "stable",
-                "retention_trend": "stable",
-                "completion_trend": "stable",
-                "trend_confidence": 0.0,
-                "significant_changes": [],
-            }
-
-            with sqlite3.connect(str(self.db.db_path)) as conn:
-                cursor = conn.cursor()
-
-                # Compare recent 7 days vs previous 14 days
-                cursor.execute(
-                    """
-                    SELECT
-                        AVG(CASE WHEN upload_date >= datetime('now', '-7 days')
-                            THEN engagement_rate END) AS recent_eng,
-                        AVG(CASE WHEN upload_date >= datetime('now', '-21 days')
-                            AND upload_date < datetime('now', '-7 days')
-                            THEN engagement_rate END) AS prior_eng,
-                        AVG(CASE WHEN upload_date >= datetime('now', '-7 days')
-                            THEN retention_rate END) AS recent_ret,
-                        AVG(CASE WHEN upload_date >= datetime('now', '-21 days')
-                            AND upload_date < datetime('now', '-7 days')
-                            THEN retention_rate END) AS prior_ret,
-                        AVG(CASE WHEN upload_date >= datetime('now', '-7 days')
-                            THEN completion_rate END) AS recent_cmp,
-                        AVG(CASE WHEN upload_date >= datetime('now', '-21 days')
-                            AND upload_date < datetime('now', '-7 days')
-                            THEN completion_rate END) AS prior_cmp,
-                        COUNT(CASE WHEN upload_date >= datetime('now', '-7 days') THEN 1 END) AS recent_count,
-                        COUNT(CASE WHEN upload_date >= datetime('now', '-21 days')
-                            AND upload_date < datetime('now', '-7 days') THEN 1 END) AS prior_count
-                    FROM video_metrics
-                    """,
-                )
-                row = cursor.fetchone()
-
-            if row and row[6] and row[6] >= 3 and row[7] and row[7] >= 3:
-                recent_eng = row[0] or 0.0
-                prior_eng = row[1] or 0.0
-                recent_ret = row[2] or 0.0
-                prior_ret = row[3] or 0.0
-                recent_cmp = row[4] or 0.0
-                prior_cmp = row[5] or 0.0
-
-                def classify(a: float, b: float) -> str:
-                    if b == 0:
-                        return "stable"
-                    pct = ((a - b) / b) * 100
-                    if pct > 5:
-                        return "increasing"
-                    if pct < -5:
-                        return "decreasing"
-                    return "stable"
-
-                trends["engagement_trend"] = classify(recent_eng, prior_eng)
-                trends["retention_trend"] = classify(recent_ret, prior_ret)
-                trends["completion_trend"] = classify(recent_cmp, prior_cmp)
-
-                sig_changes = []
-                for metric_name, recent_val, prior_val in [
-                    ("engagement_rate", recent_eng, prior_eng),
-                    ("retention_rate", recent_ret, prior_ret),
-                    ("completion_rate", recent_cmp, prior_cmp),
-                ]:
-                    if prior_val > 0:
-                        pct = ((recent_val - prior_val) / prior_val) * 100
-                        if abs(pct) > 5:
-                            sig_changes.append(
-                                {
-                                    "metric": metric_name,
-                                    "change_percent": round(pct, 1),
-                                    "time_period": "2_weeks",
-                                }
-                            )
-                trends["significant_changes"] = sig_changes
-
-                n = row[6] + row[7]
-                trends["trend_confidence"] = round(min(1.0, n / 20), 2)
-
-            return trends
-
-        except Exception as e:
-            self.logger.warning(f"Trend analysis failed: {e}")
-            return {}
-
-    def _generate_optimization_recommendations(
-        self, performance_analysis: Dict[str, Any]
-    ) -> List[OptimizationRecommendation]:
-        """Generate optimization recommendations based on performance analysis"""
-        recommendations = []
-
-        try:
-            correlations = performance_analysis.get("parameter_correlations", {})
-            enhancement_performance = performance_analysis.get(
-                "enhancement_performance", {}
-            )
-            trends = performance_analysis.get("performance_trends", {})
-
-            # Analyze each parameter for optimization opportunities
-            for param_name, constraints in self.parameter_constraints.items():
-                current_value = self._get_current_parameter_value(param_name)
-
-                if param_name in correlations:
-                    recommendation = self._analyze_parameter_optimization(
-                        param_name,
-                        current_value,
-                        correlations[param_name],
-                        constraints,
-                        enhancement_performance,
-                        trends,
-                    )
-
-                    if recommendation:
-                        recommendations.append(recommendation)
-
-            # Sort by impact estimate
-            recommendations.sort(key=lambda x: x.impact_estimate, reverse=True)
-
-        except Exception as e:
-            self.logger.error(f"Recommendation generation failed: {e}")
-
-        return recommendations
-
-    def _analyze_parameter_optimization(
-        self,
-        param_name: str,
-        current_value: float,
-        correlations: Dict[str, float],
-        constraints: Dict[str, float],
-        enhancement_performance: Dict[str, Any],
-        trends: Dict[str, Any],
-    ) -> Optional[OptimizationRecommendation]:
-        """Analyze optimization opportunity for a specific parameter"""
-        try:
-            # Calculate weighted correlation score
-            weights = {
-                "engagement_correlation": 0.4,
-                "retention_correlation": 0.4,
-                "completion_correlation": 0.2,
-            }
-            correlation_score = sum(
-                correlations.get(key, 0) * weight for key, weight in weights.items()
-            )
-
-            # Determine optimization direction
-            if correlation_score > 0.1:  # Positive correlation
-                # Should increase parameter (within constraints)
-                recommended_value = min(
-                    current_value * (1 + self.max_adjustment_per_cycle),
-                    constraints["max"],
-                )
-                direction = "increase"
-            elif correlation_score < -0.1:  # Negative correlation
-                # Should decrease parameter
-                recommended_value = max(
-                    current_value * (1 - self.max_adjustment_per_cycle),
-                    constraints["min"],
-                )
-                direction = "decrease"
-            else:
-                # No clear optimization direction
-                return None
-
-            # Snap to step size
-            step = constraints.get("step", 0.1)
-            recommended_value = round(recommended_value / step) * step
-
-            # Skip if no significant change
-            if abs(recommended_value - current_value) < step:
-                return None
-
-            # Calculate confidence based on correlation strength and sample size
-            confidence = min(
-                abs(correlation_score) * 2, 1.0
-            )  # Scale correlation to confidence
-
-            # Estimate impact
-            impact_estimate = (
-                abs(correlation_score) * 10
-            )  # Estimate percentage improvement
-
-            # Generate reason
-            reason = self._generate_optimization_reason(
-                param_name, direction, correlation_score, trends
-            )
-
-            return OptimizationRecommendation(
-                parameter_name=param_name,
-                current_value=current_value,
-                recommended_value=recommended_value,
-                confidence=confidence,
-                impact_estimate=impact_estimate,
-                reason=reason,
-            )
-
-        except Exception as e:
-            self.logger.warning(f"Parameter analysis failed for {param_name}: {e}")
-            return None
-
-    def _generate_optimization_reason(
-        self,
-        param_name: str,
-        direction: str,
-        correlation_score: float,
-        trends: Dict[str, Any],
-    ) -> str:
-        """Generate human-readable reason for optimization recommendation"""
-        correlation_strength = "strong" if abs(correlation_score) > 0.2 else "moderate"
-
-        base_reason = f"{correlation_strength.title()} positive correlation with engagement metrics"
-
-        if direction == "increase":
-            return f"{base_reason}. Increasing {param_name} should improve viewer engagement."
-        else:
-            return f"{base_reason}. Decreasing {param_name} should reduce viewer fatigue and improve retention."
-
-    def _apply_auto_optimizations(
-        self, recommendations: List[OptimizationRecommendation]
-    ) -> List[Dict[str, Any]]:
-        """Apply high-confidence recommendations automatically"""
-        applied_changes = []
-
-        try:
-            for recommendation in recommendations:
-                # Auto-apply if confidence is high enough
-                if recommendation.confidence >= self.confidence_threshold:
-                    success = self._apply_parameter_change(
-                        recommendation.parameter_name, recommendation.recommended_value
-                    )
-
-                    if success:
-                        change_record = {
-                            "parameter": recommendation.parameter_name,
-                            "old_value": recommendation.current_value,
-                            "new_value": recommendation.recommended_value,
-                            "confidence": recommendation.confidence,
-                            "estimated_impact": recommendation.impact_estimate,
-                            "reason": recommendation.reason,
-                            "applied_at": datetime.now().isoformat(),
-                        }
-                        applied_changes.append(change_record)
-
-                        self.logger.info(
-                            f"Auto-applied optimization: {recommendation.parameter_name} "
-                            f"{recommendation.current_value} -> {recommendation.recommended_value}"
-                        )
-
-        except Exception as e:
-            self.logger.error(f"Auto-optimization application failed: {e}")
-
-        return applied_changes
-
-    def _apply_parameter_change(self, parameter_name: str, new_value: float) -> bool:
-        """Apply parameter change to configuration"""
-        try:
-            # Update configuration file
-            config_updates = {parameter_name: new_value}
-            if not self._update_config_parameters(config_updates):
-                return False
-
-            # Update internal state
-            self.optimization_state["current_parameters"][parameter_name] = new_value
-
-            return True
-
-        except Exception as e:
-            self.logger.error(f"Failed to apply parameter change {parameter_name}: {e}")
-            return False
-
-    def _update_config_parameters(self, updates: Dict[str, float]) -> bool:
-        """Update configuration parameters"""
-        try:
-            if not self.config.config_file:
-                self.logger.warning(
-                    "Cannot update config parameters: no config file configured"
-                )
-                return False
-
-            config_path = Path(self.config.config_file)
-            config_data = self._load_config_yaml(config_path)
-
-            updated_paths = []
-            applied_parameters = []
-            for parameter_name, new_value in updates.items():
-                key_path = self.parameter_yaml_paths.get(parameter_name)
-                if not key_path:
-                    self.logger.warning(
-                        "Skipping unsupported optimization parameter: %s",
-                        parameter_name,
-                    )
-                    continue
-
-                self._set_nested_value(config_data, key_path, new_value)
-                updated_paths.append(".".join(key_path))
-                applied_parameters.append(parameter_name)
-
-            if not updated_paths:
-                return False
-
-            self._atomic_write_yaml(config_path, config_data)
-
-            if hasattr(self.config, "reload"):
-                self.config.reload()
-            self.config = get_config()
-
-            verified_parameters = []
-            for parameter_name in applied_parameters:
-                runtime_value = self._read_runtime_parameter_value(parameter_name)
-                expected_value = float(updates[parameter_name])
-                if runtime_value is None:
-                    continue
-                if abs(runtime_value - expected_value) > 1e-6:
-                    self.logger.warning(
-                        "Optimization parameter %s did not take effect after reload (%s != %s)",
-                        parameter_name,
-                        runtime_value,
-                        expected_value,
-                    )
-                    continue
-                verified_parameters.append(parameter_name)
-
-            if not verified_parameters:
-                return False
-
-            self.logger.info(f"Configuration updates applied: {updated_paths}")
-            return True
-
-        except Exception as e:
-            self.logger.error(f"Config update failed: {e}")
-            return False
-
-    def _load_config_yaml(self, config_path: Path) -> Dict[str, Any]:
-        """Load YAML config safely and return a dictionary root."""
-        if not config_path.exists():
-            self.logger.warning(
-                f"Config file not found, creating new mapping at {config_path}"
-            )
-            return {}
-
-        try:
-            with open(config_path, "r", encoding="utf-8") as f:
-                data = yaml.safe_load(f)
-        except yaml.YAMLError as e:
-            self.logger.error(f"Malformed YAML in {config_path}: {e}")
-            return {}
-
-        if data is None:
-            return {}
-
-        if not isinstance(data, dict):
-            self.logger.warning(
-                f"Config root is not a dict in {config_path}; resetting root structure"
-            )
-            return {}
-
-        return data
-
-    def _set_nested_value(
-        self, root: Dict[str, Any], key_path: Tuple[str, ...], value: Any
-    ):
-        """Set nested dictionary value by key path, creating dict nodes if missing."""
-        current_node = root
-        for key in key_path[:-1]:
-            child = current_node.get(key)
-            if not isinstance(child, dict):
-                child = {}
-                current_node[key] = child
-            current_node = child
-
-        current_node[key_path[-1]] = value
-
-    def _atomic_write_yaml(self, config_path: Path, yaml_data: Dict[str, Any]):
-        """Write YAML atomically using temporary file replacement."""
-        config_path.parent.mkdir(parents=True, exist_ok=True)
-
-        temp_path = None
-        try:
-            with tempfile.NamedTemporaryFile(
-                mode="w",
-                encoding="utf-8",
-                dir=str(config_path.parent),
-                prefix=f".{config_path.name}.",
-                suffix=".tmp",
-                delete=False,
-            ) as temp_file:
-                yaml.safe_dump(yaml_data, temp_file, sort_keys=False)
-                temp_path = temp_file.name
-
-            os.replace(temp_path, config_path)
-        except Exception:
-            if temp_path and os.path.exists(temp_path):
-                try:
-                    os.unlink(temp_path)
-                except OSError:
-                    pass
-            raise
-
-    def _get_current_parameter_value(self, parameter_name: str) -> float:
-        """Get current value of a parameter"""
-        runtime_value = self._read_runtime_parameter_value(parameter_name)
-        if runtime_value is not None:
-            return runtime_value
-
-        defaults = {
-            "zoom_intensity": 1.3,
-            "color_grading_strength": 0.5,
-            "background_music_volume": 0.4,
-        }
-
-        return self.optimization_state.get("current_parameters", {}).get(
-            parameter_name, defaults.get(parameter_name, 1.0)
-        )
-
-    def _read_runtime_parameter_value(self, parameter_name: str) -> Optional[float]:
-        """Read the live runtime value for supported optimization parameters."""
-        if parameter_name == "background_music_volume":
-            return float(self.config.audio.background_music_volume)
-        if parameter_name == "zoom_intensity":
-            return float(self.config.effects.max_zoom)
-        if parameter_name == "color_grading_strength":
-            return float(self.config.effects.color_grade_intensity)
-        return None
-
-    def _load_optimization_state(self) -> Dict[str, Any]:
-        """Load optimization state from storage"""
-        try:
-            state_file = self.config.paths.base_dir / "optimization_state.json"
-
-            if state_file.exists():
-                with open(state_file, "r") as f:
-                    return json.load(f)
-
-        except Exception as e:
-            self.logger.warning(f"Failed to load optimization state: {e}")
-
-        # Return default state
-        return {
-            "last_optimization": None,
-            "optimization_history": [],
-            "current_parameters": {},
-            "performance_baselines": {},
-        }
-
-    def _update_optimization_state(
-        self,
-        recommendations: List[OptimizationRecommendation],
-        applied_changes: List[Dict[str, Any]],
-    ):
-        """Update and save optimization state"""
-        try:
-            self.optimization_state["last_optimization"] = datetime.now().isoformat()
-
-            # Add to history
-            history_entry = {
-                "timestamp": datetime.now().isoformat(),
-                "recommendations_count": len(recommendations),
-                "applied_changes_count": len(applied_changes),
-                "recommendations": [asdict(r) for r in recommendations],
-                "applied_changes": applied_changes,
-            }
-
-            self.optimization_state["optimization_history"].append(history_entry)
-
-            # Keep only last 10 optimization cycles
-            self.optimization_state["optimization_history"] = self.optimization_state[
-                "optimization_history"
-            ][-10:]
-
-            # Save state
-            state_file = self.config.paths.base_dir / "optimization_state.json"
-            with open(state_file, "w") as f:
-                json.dump(self.optimization_state, f, indent=2)
-
-        except Exception as e:
-            self.logger.error(f"Failed to update optimization state: {e}")
-
-    def _generate_optimization_report(
-        self,
-        performance_analysis: Dict[str, Any],
-        recommendations: List[OptimizationRecommendation],
-        applied_changes: List[Dict[str, Any]],
-    ) -> Dict[str, Any]:
-        """Generate comprehensive optimization report"""
-        try:
-            report = {
-                "optimization_timestamp": datetime.now().isoformat(),
-                "status": "completed",
-                "analysis_summary": {
-                    "videos_analyzed": performance_analysis.get(
-                        "total_videos_analyzed", 0
-                    ),
-                    "analysis_period_days": performance_analysis.get(
-                        "analysis_period_days", 0
-                    ),
-                    "sufficient_data": performance_analysis.get(
-                        "sufficient_data", False
-                    ),
-                },
-                "recommendations": {
-                    "total_generated": len(recommendations),
-                    "high_confidence": len(
-                        [
-                            r
-                            for r in recommendations
-                            if r.confidence >= self.confidence_threshold
-                        ]
-                    ),
-                    "average_confidence": statistics.mean(
-                        [r.confidence for r in recommendations]
-                    )
-                    if recommendations
-                    else 0,
-                    "estimated_total_impact": sum(
-                        [r.impact_estimate for r in recommendations]
-                    ),
-                },
-                "applied_changes": {
-                    "total_applied": len(applied_changes),
-                    "parameters_modified": [c["parameter"] for c in applied_changes],
-                    "estimated_impact": sum(
-                        [c["estimated_impact"] for c in applied_changes]
-                    ),
-                },
-                "next_optimization": (
-                    datetime.now() + timedelta(days=self.optimization_cycle_days)
-                ).isoformat(),
-                "detailed_recommendations": [asdict(r) for r in recommendations],
-                "detailed_changes": applied_changes,
-            }
-
-            return report
-
-        except Exception as e:
-            self.logger.error(f"Report generation failed: {e}")
-            return {"status": "error", "error": str(e)}
-
-    def get_optimization_summary(self) -> Dict[str, Any]:
-        """Get summary of current optimization state"""
-        try:
-            history = self.optimization_state.get("optimization_history", [])
-
-            summary = {
-                "last_optimization": self.optimization_state.get("last_optimization"),
-                "total_optimizations": len(history),
-                "parameters_being_optimized": list(self.parameter_constraints.keys()),
-                "recent_performance": {
-                    "total_changes_last_30_days": sum(
-                        len(h["applied_changes"])
-                        for h in history[-4:]
-                        if h  # Last ~4 weeks
-                    ),
-                    "avg_confidence_last_optimization": 0,
-                    "estimated_cumulative_impact": 0,
-                },
-            }
-
-            if history:
-                last_optimization = history[-1]
-                if last_optimization.get("recommendations"):
-                    summary["recent_performance"][
-                        "avg_confidence_last_optimization"
-                    ] = statistics.mean(
-                        [r["confidence"] for r in last_optimization["recommendations"]]
-                    )
-
-                summary["recent_performance"]["estimated_cumulative_impact"] = sum(
-                    [
-                        sum([c["estimated_impact"] for c in h["applied_changes"]])
-                        for h in history[-4:]
-                    ]
-                )
-
-            return summary
-
-        except Exception as e:
-            self.logger.error(f"Summary generation failed: {e}")
-            return {}
-
-    def manual_parameter_adjustment(
-        self, parameter_name: str, new_value: float, reason: str = "Manual adjustment"
-    ) -> bool:
-        """Manually adjust a parameter with validation"""
-        try:
-            if parameter_name not in self.parameter_constraints:
-                self.logger.error(f"Unknown parameter: {parameter_name}")
-                return False
-
-            constraints = self.parameter_constraints[parameter_name]
-
-            # Validate value is within constraints
-            if new_value < constraints["min"] or new_value > constraints["max"]:
-                self.logger.error(
-                    f"Value {new_value} outside constraints for {parameter_name}"
-                )
-                return False
-
-            # Apply change
-            success = self._apply_parameter_change(parameter_name, new_value)
-
-            if success:
-                # Record manual change
-                manual_change = {
-                    "parameter": parameter_name,
-                    "old_value": self._get_current_parameter_value(parameter_name),
-                    "new_value": new_value,
-                    "type": "manual",
-                    "reason": reason,
-                    "applied_at": datetime.now().isoformat(),
-                }
-
-                history_entry = {
-                    "timestamp": datetime.now().isoformat(),
-                    "type": "manual_adjustment",
-                    "change": manual_change,
-                }
-
-                self.optimization_state["optimization_history"].append(history_entry)
-                self.logger.info(
-                    f"Manual parameter adjustment: {parameter_name} = {new_value}"
-                )
-
-            return success
-
-        except Exception as e:
-            self.logger.error(f"Manual parameter adjustment failed: {e}")
-            return False
